Squash gcc optimisations
- arm64: eliminate the .eh_frame sections from vmlinux and kernel modules
By default the aarch64 gcc generates .eh_frame sections.  Unlike
.debug_frame sections, the .eh_frame sections are loaded into memory
when the associated code is loaded.  On an example kernel being built
with this default the .eh_frame section in vmlinux used an extra 1.7MB
of memory.  The x86 disables the creation of the .eh_frame section.
The aarch64 should probably do the same to save some memory.

-  fs: Workaround the compiler's bad optimization
When compiling the kernel with  the compiler
"aarch64-linux-android-gcc (GCC) 4.9.x-google 20140827 (prerelease)"
The compiler seems to be optimizing a value out in the do_sync_write
function and is incorrectly passing the pointer instead of the
dereferenced value.
Force the compiler to use the right value by passing the dereferenced
pointer instead of another variable that is assigned with the dereferenced
value.

-  BACKPORT: arm64: ensure the kernel is compiled for LP64
commit 3d6a7b99e3fa29b92d6288487e057e0a596bd2b0 upstream.
The kernel needs to be compiled as a LP64 binary for ARM64, even when
using a compiler that defaults to code-generation for the ILP32 ABI.
Consequently, we need to explicitly pass '-mabi=lp64' (supported on
gcc-4.9 and newer).
Signed-off-by: Andrew Pinski <Andrew.Pinski@caviumnetworks.com>
Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
Reviewed-by: David Daney <ddaney@caviumnetworks.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>